Reduce diff footprint more
