+++ multisensor_pipeline/tests/test_image.py
