+++ multisensor_pipeline/tests/test_signal.py
